NLP/babycoach/requirements.txt
fastapi
uvicorn[standard]
langgraph
openai
python-dotenv
pytest

